Moved apu library from core to its own lib folder
