+++ lib/Nes_Snd_Emu/LGPL.txt
