+++ lib/Nes_Snd_Emu/nes_apu/Multi_Buffer.h
