+++ lib/Nes_Snd_Emu/nes_apu/Nonlinear_Buffer.cpp
