+++ lib/Nes_Snd_Emu/nes_apu/blargg_common.h
